requirements.txt
aiohttp
asyncio
discord.py
pyyaml
mysql
mysql-connector
dateparser
pytz
requests